refactor: split toml module into authored/processed/server
toml.rs conflated the user-authored deployment manifest with the processed
(digest-mandatory, fetched/verified) form and with the orthogonal obelisk.toml
server config. Split into submodules under config/toml/:

- authored.rs: DeploymentToml + validation and the authored *ConfigToml shapes
  (digests/component_files optional), including the webhook authored configs.
- processed.rs: resolving against the CAS and fetching/verifying into the
  runtime-ready *ConfigVerified forms (digests mandatory), plus the *Resolved-
  consuming extension traits, resolve_local_refs, and the webhook/cron verify.
- server.rs: obelisk.toml runtime config (listeners, DB, WASM globals, watchers,
  allocator, otlp/log, and the http_server declarations).
- mod.rs keeps cross-cutting helpers and re-exports the submodules so
  crate::config::toml::<Type> paths are unchanged.

The former webhook and cron submodules are dissolved: their authored, verified,
and server types are distributed by concern like every other component, and
command/server.rs uses the flattened paths. model.rs (the serde data shapes
shared by both sides, incl. the *Resolved types) stays.
